+++ packages/elements/src/Matcher/Public.ts
@@ -4,4 +4,4 @@ export type { Matcher } from "./Matcher";
 export type { ElementsMatcher } from "./ElementsMatcher";
 export type { DependenciesMatcher } from "./DependenciesMatcher";
 
-export { normalizeElementsSelector } from "./BaseElementsMatcher";
+export { normalizeElementsSelector } from "./MatcherHelpers";